pyproject.toml
[project]
name = "pcc-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb==0.5.5",
    "faker==26.0.0",
    "httpx==0.27.0",
    "langchain==0.2.11",
    "langchain-community==0.2.10",
    "langchain-openai==0.1.17",
    "langgraph==0.1.19",
    "numpy==1.26.4",
    "openai==1.35.13",
    "pandas==2.2.2",
    "psycopg2-binary==2.9.9",
    "pydantic==2.8.2",
    "pydantic-settings==2.4.0",
    "pypdf==4.3.1",
    "pypdf2==3.0.1",
    "pytest==8.3.2",
    "pytest-asyncio==0.23.8",
    "pytest-cov==5.0.0",
    "python-docx==1.1.2",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "sentence-transformers==3.0.1",
    "sqlalchemy==2.0.31",
    "structlog==24.4.0",
    "tiktoken==0.7.0",
]
